requirements.txt
absl-py==0.8.1
apex==0.1
arrow==0.14.2
astor==0.8.0
Babel==2.8.0
beautifulsoup4==4.7.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
binaryornot==0.4.4
blis==0.4.1
boto3==1.10.14
botocore==1.13.14
cachetools==3.1.1
catalogue==1.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
croniter==0.3.31
cycler==0.10.0
cymem==2.0.2
dask==2.9.1
distributed==2.9.1
docker==4.1.0
docutils==0.15.2
et-xmlfile==1.0.1
fsspec==0.6.2
ftfy==5.6
future==0.17.1
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
GPUtil==1.4.0
grpcio==1.25.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
locket==0.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-oneofschema==2.0.1
matplotlib==3.1.3
mongoengine==0.18.2
more-itertools==7.2.0
msgpack==0.6.2
murmurhash==1.0.2
mypy-extensions==0.4.3
nltk==3.4.5
numpy==1.17.2
oauthlib==3.1.0
openpyxl==3.0.1
opt-einsum==3.1.0
pandas==0.25.1
partd==1.1.0
pendulum==2.0.5
Pillow==7.0.0
plac==1.1.3
plotly==4.5.0
poyo==0.4.2
prefect==0.8.1
preshed==3.0.2
protobuf==3.10.0
psutil==5.6.7
PTable==0.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pymongo==3.9.0
PyMySQL==0.9.3
pyparsing==2.4.6
python-box==3.4.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.11.2
python-slugify==4.0.0
python-socketio==4.4.0
pytorch-ignite==0.2.1
pytorch-transformers==1.2.0
pytz==2019.2
pytzdata==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
redis==3.4.1
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.2.2
rsa==4.0
s3transfer==0.2.1
sacremoses==0.0.35
scikit-learn==0.21.3
scipy==1.3.2
sentencepiece==0.1.83
seqeval==0.0.12
simpletransformers==0.6.4
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.2
spacy==2.2.2
srsly==0.2.0
supervisor==4.1.0
tabulate==0.8.6
tblib==1.6.0
tensorboard==1.15.0
tensorboardX==1.9
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
text-unidecode==1.3
thinc==7.3.1
toml==0.10.0
toolz==0.10.0
torch==1.3.1
tornado==6.0.3
tqdm==4.38.0
transformers==2.1.1
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.3
wasabi==0.4.0
wcwidth==0.1.7
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
whichcraft==0.6.0
wordcloud==1.6.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.6
zict==1.0.0
zipp==0.6.0
